First stab at 32 bit ints, changed timers to use a difference in time rather than an absolute time. Seems to work - BUT OneWire addresses need changing to strings, and JsNative/wrapper stuff could be simplified now that JsVarInt==int32

+++ src/jsutils.h
@@ -22,6 +22,7 @@
 #include <stdlib.h>
 #endif
 #include <stdarg.h> // for va_args
+#include <stdint.h>
 
 #ifdef LINUX
 #include <math.h>
@@ -98,8 +99,8 @@ typedef enum {FALSE = 0, TRUE = !FALSE} bool;
   #endif
 #endif
 
-typedef long long JsVarInt;
-typedef unsigned long long JsVarIntUnsigned;
+typedef int32_t JsVarInt;
+typedef uint32_t JsVarIntUnsigned;
 #ifdef USE_FLOATS
 typedef float JsVarFloat;
 #else
